webapp/backend/requirements.txt
fastapi[all]
python-jose[cryptography]
passlib
psycopg2-binary

ijson
